+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Nothing
 
 ### Fixed
-- Nothing
+- Fixed bug where the dynamic voxel save file browser doesn't appear in Recast.Demo
 
 ### Changed
 - Changed to reuse samples and edges list in BuildPolyDetail()